restructure: move `AttestationRecord` to storage
